backend_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic==2.10.5
pydantic-core==2.27.2
modal>=0.0.95
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
python-dotenv>=0.19.0
wordcloud>=1.8.0
pyLDAvis>=3.3.1
gensim>=4.0.0
nomic==3.4.1 